pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "complexity-cli"
version = "0.1.0"
description = "CLI tool to analyze GitHub PR complexity using LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Complexity Analyzer"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "boto3>=1.34.0",
    "tiktoken>=0.7.0",
    "typer>=0.9.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
complexity-cli = "cli.main:app"

[tool.setuptools.packages.find]
include = ["cli*", "reports*"]

[tool.setuptools.package-data]
cli = ["prompt/*.txt"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

